+++ src/scandeval/benchmark_dataset.py
@@ -633,7 +633,7 @@ class BenchmarkDataset(ABC):
 
     @abstractmethod
     def _apply_few_shot_prompt(
-        self, examples: dict, few_shot_examples: list[dict]
+        self, examples: dict, few_shot_examples: list[dict], tokenizer: Tokenizer
     ) -> dict:
         """Apply a few-shot prompt to the examples.
 
@@ -642,6 +642,9 @@ class BenchmarkDataset(ABC):
                 The examples to apply the prompt to.
             few_shot_examples:
                 The examples to be included in the few-shot prompt.
+            tokenizer:
+                The tokenizer to use to tokenise the examples. Used to apply the
+                chat template if the model we're benchmarking is instruction tuned.
 
         Returns:
             The examples with the few-shot prompt applied.