+++ src/scandeval/sequence_classification.py
@@ -229,7 +229,7 @@ class SequenceClassification(BenchmarkDataset):
         return few_shot_examples
 
     def _apply_few_shot_prompt(
-        self, examples: dict, few_shot_examples: list[dict]
+        self, examples: dict, few_shot_examples: list[dict], tokenizer: Tokenizer
     ) -> dict:
         """Apply a few-shot prompt to the examples.
 
@@ -238,6 +238,9 @@ class SequenceClassification(BenchmarkDataset):
                 The examples to apply the prompt to.
             few_shot_examples:
                 The examples to be included in the few-shot prompt.
+            tokenizer:
+                The tokenizer to use to tokenise the examples. Used to apply the
+                chat template if the model we're benchmarking is instruction tuned.
 
         Returns:
             The examples with the few-shot prompt applied.
@@ -254,7 +257,6 @@ class SequenceClassification(BenchmarkDataset):
         prompt_prefix = ""
         if self.dataset_config.prompt_prefix:
             prompt_prefix = self.dataset_config.prompt_prefix + "\n\n"
-        few_shot_prompt = prompt_prefix + "\n\n".join(few_shot_prompts)
 
         # Add the texts from the examples to the prompts. We remove newlines from the
         # examples as they have the special function to separate the few-shot examples
@@ -266,10 +268,20 @@ class SequenceClassification(BenchmarkDataset):
             for text in examples["text"]
         ]
 
-        examples["text"] = [
+        few_shot_prompt = prompt_prefix + "\n\n".join(few_shot_prompts)
+        final_prompts = [
             few_shot_prompt + "\n\n" + new_prompt for new_prompt in new_prompts
         ]
 
+        if tokenizer.chat_template is not None:
+            final_prompts = [
+                tokenizer.apply_chat_template(
+                    conversation=[dict(role="user", content=prompt)],
+                )
+                for prompt in final_prompts
+            ]
+
+        examples["text"] = final_prompts
         return examples
 
     def _extract_labels_from_generation(