+++ src/scandeval/types.py
@@ -7,6 +7,7 @@ import numpy as np
 ScoreDict = dict[str, dict[str, float] | dict[str, list[dict[str, float]]]]
 Predictions = np.ndarray | list[str] | list[list[str]]
 Labels = np.ndarray | list[str] | list[list[str]]
+Conversation = list[dict[str, str]]
 
 
 def is_list_of_int(x: Any) -> TypeGuard[list[int]]: